refactor: extract csvReporter
