+++ src/reporters/humanReporter.ts
@@ -9,7 +9,7 @@ import chalk from 'chalk';
 import { get, getArray, isPlainObject, isString, Optional } from '@salesforce/ts-types';
 import { Messages } from '@salesforce/core';
 import { Field, FieldType, SoqlQueryResult } from '../dataSoqlQueryTypes.js';
-import { QueryReporter, logFields, isSubquery, isAggregate, getAggregateAliasOrName } from '../reporters.js';
+import { QueryReporter, logFields, isSubquery, isAggregate, getAggregateAliasOrName } from './reporters.js';
 
 type ParsedFields = {
   /** Field names */
